+++ test/test_map_DXM.py
@@ -1,3 +1,4 @@
+import glob
 import os
 import shutil
 from pathlib import Path
@@ -35,11 +36,9 @@ with initialize(version_base="1.2", config_path="../configs"):
         ],
     )
 
-EXPECTED_DTM_INTERP = os.path.join("DTM", "test_data_77055_627760_LA93_IGN69_interp.tif")
-EXPECTED_DTM_HILLSHADE = os.path.join("tmp_dtm", "hillshade", "test_data_77055_627760_LA93_IGN69_hillshade.tif")
 EXPECTED_DTM_COLORED = [
-    os.path.join("DTM", "color", "1cycle", "test_data_77055_627760_LA93_IGN69_DTM_hillshade_color1c.tif"),
-    os.path.join("DTM", "color", "4cycles", "test_data_77055_627760_LA93_IGN69_DTM_hillshade_color4c.tif"),
+    os.path.join("DTM_FINAL", "1cycle", "test_data_77055_627760_LA93_IGN69_DTM_hillshade_color1c.tif"),
+    os.path.join("DTM_FINAL", "4cycles", "test_data_77055_627760_LA93_IGN69_DTM_hillshade_color4c.tif"),
 ]
 
 
@@ -72,7 +71,6 @@ def test_create_raw_DXM_default_pixelsize():
     output_dir = os.path.join(OUTPUT_DIR, "create_raw_DXM_default_pixelsize")
     raster_dxm_raw = os.path.join(
         output_dir,
-        CONFIG_2.dtm.intermediate_dirs.dxm_raw,
         f"{TILENAME}_interp{CONFIG_2.io.extension}",
     )
     os.makedirs(os.path.dirname(raster_dxm_raw))
@@ -85,27 +83,78 @@ def test_create_raw_DXM_default_pixelsize():
         config_io=CONFIG_2.io,
     )
 
-    assert os.path.isfile(os.path.join(output_dir, EXPECTED_DTM_INTERP))
-    assert raster_dxm_raw == os.path.join(output_dir, EXPECTED_DTM_INTERP)
+    assert os.path.isfile(raster_dxm_raw)
+    with rasterio.open(raster_dxm_raw) as raster:
+        assert raster.res == (1, 1)
+        assert raster.bounds == (
+            COORDX * TILE_COORD_SCALE - 0.5,
+            COORDY * TILE_COORD_SCALE - TILE_WIDTH + 0.5,
+            COORDX * TILE_COORD_SCALE + TILE_WIDTH - 0.5,
+            COORDY * TILE_COORD_SCALE + 0.5,
+        )
 
 
-def test_create_colored_dxm_with_hillshade_dtm_1m():
+def test_create_colored_dxm_with_hillshade_dtm_1m_default():
     """
     Verify :
         - all .tif are created
         - pixel size is 1m
     """
-    output_dir = os.path.join(OUTPUT_DIR, "create_colored_dxm_with_hillshade_dtm_1m")
-    conf_dtm = CONFIG_2.dtm.copy()
-    conf_dtm.color.cycles_DTM_colored = [1, 4]
-    conf_io = CONFIG_2.io.copy()
-    conf_io.output_dir = output_dir
+    output_dir = os.path.join(OUTPUT_DIR, "create_colored_dxm_with_hillshade_dtm_1m_default")
+    with initialize(version_base="1.2", config_path="../configs"):
+        cfg = compose(
+            config_name="config_ctview",
+            overrides=[
+                f"io.output_dir={output_dir}",
+                f"io.tile_geometry.tile_coord_scale={TILE_COORD_SCALE}",
+                f"io.tile_geometry.tile_width={TILE_WIDTH}",
+                "dtm.color.cycles_DTM_colored=[1,4]",
+            ],
+        )
     map_DXM.create_colored_dxm_with_hillshade(
-        input_las=INPUT_FILE, tilename=TILENAME, config_dtm=conf_dtm, config_io=conf_io
+        input_las=INPUT_FILE, tilename=TILENAME, config_dtm=cfg.dtm, config_io=cfg.io
     )
 
-    assert os.path.isfile(os.path.join(output_dir, EXPECTED_DTM_INTERP))
-    assert os.path.isfile(os.path.join(output_dir, EXPECTED_DTM_HILLSHADE))
+    assert os.listdir(output_dir) == ["DTM_FINAL"]
+    assert not glob.glob("tmp/tmp_dxm*")
+
+    for p in EXPECTED_DTM_COLORED:
+        assert os.path.isfile(os.path.join(output_dir, p))
+        with rasterio.open(os.path.join(output_dir, p)) as raster:
+            assert raster.res == (1, 1)
+
+
+def test_create_colored_dxm_with_hillshade_dtm_1m_and_intermediate_files():
+    """
+    Verify :
+        - all .tif are created
+        - pixel size is 1m
+    """
+    output_dir = os.path.join(OUTPUT_DIR, "create_colored_dxm_with_hillshade_dtm_1m_and_intermediate_files")
+    with initialize(version_base="1.2", config_path="../configs"):
+        cfg = compose(
+            config_name="config_ctview",
+            overrides=[
+                f"io.output_dir={output_dir}",
+                f"io.tile_geometry.tile_coord_scale={TILE_COORD_SCALE}",
+                f"io.tile_geometry.tile_width={TILE_WIDTH}",
+                "dtm.intermediate_dirs.dxm_raw=DTM_VAL",
+                "dtm.intermediate_dirs.dxm_hillshade=tmp_dtm/hillshade",
+                "dtm.intermediate_dirs.folder_LUT=DTM_FINAL/LUT",
+                "dtm.color.cycles_DTM_colored=[1,4]",
+            ],
+        )
+    map_DXM.create_colored_dxm_with_hillshade(
+        input_las=INPUT_FILE, tilename=TILENAME, config_dtm=cfg.dtm, config_io=cfg.io
+    )
+    expected_dtm_interp = os.path.join("DTM_VAL", "test_data_77055_627760_LA93_IGN69_interp.tif")
+    expected_dtm_hillshade = os.path.join("tmp_dtm", "hillshade", "test_data_77055_627760_LA93_IGN69_hillshade.tif")
+    expected_LUT_1c = os.path.join("DTM_FINAL/LUT", "LUT_1cycle_test_data_77055_627760_LA93_IGN69.txt")
+    expected_LUT_4c = os.path.join("DTM_FINAL/LUT", "LUT_4cycle_test_data_77055_627760_LA93_IGN69.txt")
+    assert os.path.isfile(os.path.join(output_dir, expected_dtm_interp))
+    assert os.path.isfile(os.path.join(output_dir, expected_dtm_hillshade))
+    assert os.path.isfile(os.path.join(output_dir, expected_LUT_1c))
+    assert os.path.isfile(os.path.join(output_dir, expected_LUT_4c))
     for p in EXPECTED_DTM_COLORED:
         assert os.path.isfile(os.path.join(output_dir, p))
         with rasterio.open(os.path.join(output_dir, p)) as raster:
@@ -165,21 +214,9 @@ def test_add_dxm_hillshade_to_raster_density_5m():
     """
     output_dir = os.path.join(OUTPUT_DIR, "add_dxm_hillshade_to_raster_density_5m")
     input_raster = os.path.join(INPUT_DIR_RASTER, f"test_data_{COORDX}_{COORDY}_LA93_IGN69_DENS_COLOR_5m.tif")
-    output_raster = os.path.join(
-        output_dir,
-        CONFIG_2.density.output_subdir,
-        f"{TILENAME}_DENS{CONFIG_2.io.extension}",
-    )
-    raster_dxm_raw = os.path.join(
-        output_dir,
-        CONFIG_2.density.intermediate_dirs.dxm_raw,
-        f"{TILENAME}_interp{CONFIG_2.io.extension}",
-    )
-    raster_dxm_hillshade = os.path.join(
-        output_dir,
-        CONFIG_2.density.intermediate_dirs.dxm_hillshade,
-        f"{TILENAME}_hillshade{CONFIG_2.io.extension}",
-    )
+    output_raster = os.path.join(output_dir, CONFIG_2.density.output_subdir, f"{TILENAME}_DENS{CONFIG_2.io.extension}")
+    raster_dxm_raw = os.path.join(output_dir, f"{TILENAME}_interp{CONFIG_2.io.extension}")
+    raster_dxm_hillshade = os.path.join(output_dir, f"{TILENAME}_hillshade{CONFIG_2.io.extension}")
     map_DXM.add_dxm_hillshade_to_raster(
         input_raster=input_raster,
         input_pointcloud=INPUT_FILE,
@@ -193,10 +230,9 @@ def test_add_dxm_hillshade_to_raster_density_5m():
         config_io=CONFIG_2.io,
     )
 
-    assert os.path.isfile(os.path.join(output_dir, EXPECTED_DTM_DENS_INTERP))
-    assert os.path.isfile(os.path.join(output_dir, EXPECTED_DTM_DENS_HILLSHADE))
+    assert os.path.isfile(raster_dxm_raw)
+    assert os.path.isfile(raster_dxm_hillshade)
     assert output_raster == os.path.join(output_dir, EXPECTED_DENS_WITH_HILLSHADE)
-    assert os.path.isfile(os.path.join(output_dir, EXPECTED_DENS_WITH_HILLSHADE))
     with rasterio.open(output_raster) as raster:
         assert raster.res == (5, 5)
 
@@ -239,7 +275,7 @@ def test_create_colored_dxm_with_hillshade():
     map_DXM.create_colored_dxm_with_hillshade(str(las_with_buffer), input_tilename, cfg.dtm, cfg.io)
     with rasterio.Env():
         with rasterio.open(
-            Path(output_dir) / "DTM" / "color" / "1cycle" / f"{input_tilename}_DTM_hillshade_color1c.tif"
+            Path(output_dir) / "DTM_FINAL" / "1cycle" / f"{input_tilename}_DTM_hillshade_color1c.tif"
         ) as raster:
             band1 = raster.read(1)
             band2 = raster.read(2)
@@ -249,7 +285,7 @@ def test_create_colored_dxm_with_hillshade():
             assert band3[8, 8] == 0
             assert raster.res == (1, 1)
         with rasterio.open(
-            Path(output_dir) / "DTM" / "color" / "4cycles" / f"{input_tilename}_DTM_hillshade_color4c.tif"
+            Path(output_dir) / "DTM_FINAL" / "4cycles" / f"{input_tilename}_DTM_hillshade_color4c.tif"
         ) as raster:
             band1 = raster.read(1)
             band2 = raster.read(2)